### 2019-12-20 #1976 @whedon
[PRE REVIEW]: Recan: Python tool for detection of recombination events in viral genomes
**Submitting author:** @babinyurii (<a href="http://orcid.org/0000-0002-7524-5921">Yuriy Babin</a>)
**Repository:** <a href="https://github.com/babinyurii/recan" target ="_blank">https://github.com/babinyurii/recan</a>
**Version:** v0.1
