### 2019-12-20 8db6d22
Update paper.md
