### 2019-12-20 cc9599b
Update paper.md
