+++ paper/paper.md
@@ -22,12 +22,12 @@ Recombination is widespread in viruses and is considered to  be one of the mecha
 The similarity plot based on genetic distance between nucleotide sequences is an intuitive and simple way to explore and discover recombination events in viral genomes. 
 Python package named `recan` (recombination analyzer) allows to construct genetic distance plots and explore them interactively.  Recan is intended to be used via Jupyter notebook which is a popular instrument for data analysis.
 The aim of the package is to give the researchers who are familiar with Python a tool which allows to perform recombination analysis easily and quickly.
-The advantages of `recan` above the distance methods implemented previously in desktop software [@Lole1999; @Etherington2005a; @Martin2015a] are its interactivity, speed, the ability to easily include sequences into or exclude them from the output without reconstructing the plot and recalculating the distance data, and, at last, the ability to analyze simultaneously several datasets during a single session.
+The advantages of `recan` above the distance methods implemented previously in desktop software [@Lole1999; @Etherington2005a; @Martin2015] are its interactivity, speed, the ability to easily include sequences into or exclude them from the output without reconstructing the plot and recalculating the distance data, and, at last, the ability to analyze simultaneously several datasets during a single session.
 Recan is based on Biopython, Pandas, and Plotly libraries. 
 The package requires sequence alignment in fasta format as an input. A user can control the sliding window size, the window shift, a method of distance calculation, a sequence of interest (which is a sequence potentially containing recombination region), and the alignment region which will be included into distance calculation. Two methods of genetic distance calculation implemented so far are pairwise and Kimura 2-parameter distance.  After a recombination event is found, the distance data can be saved in csv or excel  file, or directly used to reconstruct the plot in Jupyter notebook with ones favorite plotting library to get a final report.
 
 # Testing and verification
-To test the package four previously reported recombinant viral genomes were used: human immunodeficiency virus (HIV) [@Liitsola2000a], hepatitis C virus (HCV) [@Smith2014a] norovirus [@Jiang1999a], and lumpy skin disease virus (LSDV) [@Sprygin2018].
+To test the package four previously reported recombinant viral genomes were used: human immunodeficiency virus (HIV) [@Liitsola2000a], hepatitis C virus (HCV) [@Smith2014] norovirus [@Jiang1999], and lumpy skin disease virus (LSDV) [@Sprygin2018].
 Each dataset included a recombinant sequence, its parental sequences and a number of other sequences of the same virus closely related to the recombinant.
 HIV, HCV and Norovirus sequences were aligned with Clustal W [@Larkin2007] and LSDV genomes were aligned with MAFFT [@Katoh2002] by using Ugene software [@Okonechnikov2012].
 Finally, HIV alignment included 25 sequences of the length 3135 bp, HCV alignment had 23 sequences of the length 9431, Norovirus alignment included 19 genomes of the length 3366, and LSDV dataset had 14 sequences of the length 150511 bp. 
