+++ paper/paper.md
@@ -39,19 +39,19 @@ Distance plots with recombination events detected by recan in the test datasets 
 Recan can be installed by pip Python package manager using `pip install recan` command. The source code, guide and datasets are available at https://github.com/babinyurii/recan. Recan is supported on Linux and Windows.
 
 
-![](/pictures/hiv_rec_kal153.png)
+![](/paper_plots/hiv_rec_kal153.png)
 _Figure 1. HIV recombinant strain AF193276 between sequences AF193275 and AF193278._
 
 
-![](/pictures/hcv_2k_1b_rec.png)
+![](/paper_plots/hcv_2k_1b_rec.png)
 _Figure 2. HCV intergenotype recombinant 2k/1b._
 
 
-![](/pictures/norovirus_rec.png)
+![](/paper_plots/norovirus_rec.png)
 _Figure 3. Norovirus  recombinant AF190817 between parental sequences U22498 and X86557._
 
 
-![](/pictures/lsdv_rec_sar.png)
+![](/paper_plots/lsdv_rec_sar.png)
 _Figure 4. LSDV recombinant vaccine-like strain LSDV RUSSIA/Saratov/2017 between sequences AF193275 and KY829023._
 
 
