### 2019-12-20 f1cc7ea
Update paper.md
