+++ paper/paper.md
@@ -20,9 +20,9 @@ bibliography: paper.bib
 
 Recombination is widespread in viruses and is considered to  be one of the mechanisms that drives evolutionary changes and helps viruses to overcome selective pressure and adapt to new environments [@Perez-Losada2015].
 The similarity plot based on genetic distance between nucleotide sequences is an intuitive and simple way to explore and discover recombination events in viral genomes. 
-Python package named `recan` (recombination analyzer) allows to construct genetic distance plots and explore them interactively.  Recan is intended to be used via Jupyter notebook which is a popular instrument for data analysis.
+Python package named recan (recombination analyzer) allows to construct genetic distance plots and explore them interactively.  Recan is intended to be used via Jupyter notebook which is a popular instrument for data analysis.
 The aim of the package is to give the researchers who are familiar with Python a tool which allows to perform recombination analysis easily and quickly.
-The advantages of `recan` above the distance methods implemented previously in desktop software [@Lole1999; @Etherington2005a; @Martin2015] are its interactivity, speed, the ability to easily include sequences into or exclude them from the output without reconstructing the plot and recalculating the distance data, and, at last, the ability to analyze simultaneously several datasets during a single session.
+The advantages of recan above the distance methods implemented previously in desktop software [@Lole1999; @Etherington2005a; @Martin2015] are its interactivity, speed, the ability to easily include sequences into or exclude them from the output without reconstructing the plot and recalculating the distance data, and, at last, the ability to analyze simultaneously several datasets during a single session.
 Recan is based on Biopython, Pandas, and Plotly libraries. 
 The package requires sequence alignment in fasta format as an input. A user can control the sliding window size, the window shift, a method of distance calculation, a sequence of interest (which is a sequence potentially containing recombination region), and the alignment region which will be included into distance calculation. Two methods of genetic distance calculation implemented so far are pairwise and Kimura 2-parameter distance.  After a recombination event is found, the distance data can be saved in csv or excel  file, or directly used to reconstruct the plot in Jupyter notebook with ones favorite plotting library to get a final report.
 
