### 2019-12-20 #1976 @labarba
@will-rowe — Could you handle this submission as editor?
Have look please. We understand of course that these days will be slow and you might be on holiday. I also realize that I have asked you to take another submission... we have received half a dozen submissions in various flavors of bioinformatics, and our all editors in this topic have been asked!
