+++ paper/paper.md
@@ -9,7 +9,7 @@ authors:
   orcid: 0000-0002-7524-5921
   affiliation: "1"
 affiliations:
-- name: viral hepatitis laboratory, Central Research Institute of Epidemiology, Moscow, Russia
+- name: 
   index: 1
 date: 29 November 2019
 bibliography: paper.bib
