### 2020-01-09 #1976 @JoaoRodrigues
Hi @will-rowe

Thanks for the invitation but I am not sure I have the right expertise to review this paper adequately, particuarly in terms of relevance to the field and appropriate methodology. If you're fine with me reviewing only the software component, I will be more than happy to contribute.

Cheers
