### 2020-01-09 #1976 @will-rowe
Hey @flashton2003 - would you be willing to review this paper for JOSS?
