### 2020-01-09 #1976 @rhagenson
@will-rowe I seem to be in the same boat as João in that I lack the necessary expertise to review the paper adequately. I do not work with viral genomes, nor know the solution space of similar tools to judge this tool's merit within appropriate context; therefore, I must politely decline the request to review.
