### 2020-01-10 #1976 @will-rowe
Hey @flashton2003 - I was hoping you might know someone if you couldn't do it. Thanks very much for the tip off.

Hi @lamhm - it would be fantastic if you could review this submission for us. Please do let me know if you can or if you need any information on how the review process works at JOSS.
