### 2020-01-14 ec102a7
Update paper.md
