+++ paper/paper.md
@@ -8,8 +8,8 @@ authors:
 - name: Yuriy Babin
   orcid: 0000-0002-7524-5921
   affiliation: "1"
-affiliations: National Medical Research Center for Tuberculosis and Infectious Diseases
-- name: 
+affiliations: 
+- name: National Medical Research Center for Tuberculosis and Infectious Diseases
   index: 1
 date: 29 November 2019
 bibliography: paper.bib
