### 2020-01-14 #1976 @lamhm
Hi @will-rowe,

Sorry, I did not notice that I had a message on Github. I am happy to help. But I do not familiar with this style of review. Do you have a guideline that I can follow?
