### 2020-01-14 #1976 @will-rowe
Fantastic - thank you very much @lamhm!

Reviewer criteria and checklist for completing a review can be found [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Once I have confirmed a second reviewer, we will start the review process officially - this is when we create a new GitHub issue where we can keep track of the reviews etc. This new issue will be where you complete the reviewer checklist and where we interact (in terms of asking for revisions, confirming acceptance. If you have any revisions/edits for the author, it's best to make those as new issues on the repository of the software being reviewed, then reference the issue in the review issue if needed.

Hope this makes sense to you, let me know if you have any queries and I'll let you know when we're moving this into review (but you can start your review now if you like and complete the checklist in due course).
