### 2020-01-14 #1976 @will-rowe
Hi @betteridiot - would you be able to review this paper for JOSS?
