### 2020-01-14 #2014 @whedon
[REVIEW]: Recan: Python tool for detection of recombination events in viral genomes
**Submitting author:** @babinyurii (<a href="http://orcid.org/0000-0002-7524-5921">Yuriy Babin</a>)
**Repository:** <a href="https://github.com/babinyurii/recan" target ="_blank">https://github.com/babinyurii/recan</a>
**Version:** 0.1.3
