### 2020-01-28 #2014 @betteridiot
Thank you @will-rowe for pinging us. I am still actively reviewing the paper, but I do have a question regarding the review process. Regarding the "Quality of writing" stage; are we expected to give revisions or specific notes or just request them to have an editor review it for them?
