### 2020-01-28 babinyurii/recan#3 @betteridiot
JoSS Review: Community Guidelines
In conjunction with the review of your package to JoSS (available at https://github.com/openjournals/joss-reviews/issues/2014), here is an issue for you to address for your submission.

In the spirit of Open Source, programs submitted to JoSS should iterate on the concept of developing a Community of Practice. Therefore, most submissions should contain a set of Community Guidelines. These guidelines should detail how to contribute, how to report issues, and how to seek support.

These can easily be detailed in a CONTRIBUTING.md file at the top-level of the package and consider setting up [Issue/PR templates](https://help.github.com/en/github/building-a-strong-community/about-issue-and-pull-request-templates) for your repository.
