### 2020-01-28 babinyurii/recan#4 @betteridiot
JoSS Review: Installation
In conjunction with the review of your package to JoSS (available at https://github.com/openjournals/joss-reviews/issues/2014), here is an issue for you to address for your submission.

As the community migrates/upgrades to using Jupyter Lab, you may want to consider including instructions on how to install the Jupyter Lab Plotly renderer: https://plot.ly/python/getting-started/#jupyterlab-support-python-35 to meet your package's visualization needs.
