### 2020-01-28 babinyurii/recan#5 @betteridiot
JoSS Review: README Documentation
In conjunction with the review of your package to JoSS (available at https://github.com/openjournals/joss-reviews/issues/2014), here is an issue for you to address for your submission.

While your GitHub repo contains a README file, this file is not included in your package on PyPI. As it stands, any person looking for your package through that venue will not be able to view your package details unless they were to navigate to the repo itself. Consider packaging your README with your PyPI package: https://packaging.python.org/guides/making-a-pypi-friendly-readme/
