### 2020-01-28 babinyurii/recan#6 @betteridiot
JoSS Review: Example Usage
In conjunction with the review of your package to JoSS (available at https://github.com/openjournals/joss-reviews/issues/2014), here is an issue for you to address for your submission.

While I appreciate the wealth of examples that were added to the documentation, none of these examples explain what is being visualized thoroughly. A lay-person, or early scientist in the field, may look at these plots and not understand _where_ the events are occurring and _what_ they look like. Consider adding brief annotations to the screen shots explaining what/where these events are.
