### 2020-01-28 babinyurii/recan#7 @betteridiot
JoSS Review: Automated tests
In conjunction with the review of your package to JoSS (available at https://github.com/openjournals/joss-reviews/issues/2014), here is an issue for you to address for your submission.

Regarding automated testing, I appreciate that you already have some in place. However, there are some points to be made:
1. You do not detail how to run the automated tests in the instructions/README
1. One of the industry leaders in running automated tests is [pytest](https://docs.pytest.org/en/latest/), and if a naive user were to attempt to run your test suite using it, the naming conventions you have in place for the test folder and files would inhibit test discovery:
    * Consider renaming `test/` to `tests/` and `test.py` to `recan_test.py` or `basic_test.py`
1. Your tests require the user to run tests from the `test/` folder itself. Running the automated tests from the top-level directory causes tests to fail. If this behavior is desired, please detail it in the instructions for running the tests.
1. Lastly, with regard to test coverage:
```shell
$ pytest --cov=recan .
================================================= test session starts ==================================================platform linux -- Python 3.6.7, pytest-5.3.2, py-1.8.1, pluggy-0.12.0
rootdir: /.../recan/recan
plugins: cov-2.8.1
collected 13 items

recan_test.py .............                                                                                      [100%]
=================================================== warnings summary ===================================================/.../miniconda3/envs/recan/lib/python3.6/site-packages/Bio/Alphabet/__init__.py:26
  /.../miniconda3/envs/recan/lib/python3.6/site-packages/Bio/Alphabet/__init__.py:26: PendingDeprecationWarning:

  We intend to remove or replace Bio.Alphabet in 2020, ideally avoid using it explicitly in your code. Please get in touch if you will be adversely affected by this. https://github.com/biopython/biopython/issues/2046

-- Docs: https://docs.pytest.org/en/latest/warnings.html

----------- coverage: platform linux, python 3.6.7-final-0 -----------
Name                                  Stmts   Miss  Cover
---------------------------------------------------------
/.../recan/recan/recan/__init__.py       0      0   100%
/.../recan/recan/recan/simgen.py       147     48    67%
---------------------------------------------------------
TOTAL                                   147     48    67%

============================================ 13 passed, 1 warning in 3.23s ============================================= 11:32
```
