### 2020-01-28 babinyurii/recan#9 @betteridiot
JoSS Review: Statement of need
In conjunction with the review of your package to JoSS (available at https://github.com/openjournals/joss-reviews/issues/2014), here is an issue for you to address for your submission.

My only comment here is that a clear problem that `recan` solves. I do see that it is meant for researchers that use Python and that it has some advantages over RDP4 and RAT. However, a concise statement of need seems absent.
