### 2020-01-28 babinyurii/recan#10 @betteridiot
JoSS Review: State of the field
In conjunction with the review of your package to JoSS (available at https://github.com/openjournals/joss-reviews/issues/2014), here is an issue for you to address for your submission.

While `recan` allows the plotting of genetic distances (much like RDP4 and RAT), and it does so interactively and ad hoc, it foregoes the ability to detect potential recombination events. This is one of the primary reasons others use those programs.

Are there any future plans to implement potential recombination event detection into `recan`?
