### 2020-01-28 babinyurii/recan#11 @betteridiot
JoSS Review: Quality of writing
In conjunction with the review of your package to JoSS (available at https://github.com/openjournals/joss-reviews/issues/2014), here is an issue for you to address for your submission.

This is potentially the most difficult issue to bring up. While the implementation of `recan` is tremendous work, the readability of this submission could use significant revisions.

I planned on making a series of revision notes to pass on to the author but, after looking at the list, I will refrain as I know what it feels like to have a giant list of revisions to come from a reviewer.

Generally speaking, this manuscript would benefit from improvements in grammar and sentence structure.
