### 2020-01-28 #2014 @betteridiot
@will-rowe, @babinyurii these are my reviewer comments for the review of `recan` as the package stands right now.
