### 2020-01-29 babinyurii/recan#10 @ghost
I agree with the comment of @betteridiot above. I also think this is the biggest weakness of the `recan`. While the package offers a very good visualisation tool for analysing genetic similarity, I would not call it a recombination detection tool since the detection requires completely manual assessment of the similarity plots.

In modern contexts, where we often have to work with a lot of data at once, I think it would be extremely helpful if `recan` can incorporate a method to detect potential recombination events.
