### 2020-01-29 babinyurii/recan#12 @lamhm
JoSS Review: Performance
In conjunction with the review of your package to JoSS (available at [openjournals/joss-reviews#2014](https://github.com/openjournals/joss-reviews/issues/2014)), here is an issue for you to address for your submission.

In your paper, you have stated that one of the advantages of `recan` over other methods of recombination detection (e.g. RAT and RDP4) is **speed**. However, I do not think the speed of `recan` should be compared to that of the aforementioned programs as the `recombination auto-search function` (which consumes a lot of run-time and is the main focus of RAT and RDP4) has not been implemented in `recan`.
