### 2020-01-29 babinyurii/recan#6 @ghost
The author should also specify how users can access the example datasets. Otherwise, the following line of code would be confusing for inexperienced users: `sim_obj = Simgen("./data/hbv_C_Bj_Ba.fasta")`.

On a side note, I guess the author have changed the name of the `data` folder in their Github repository into `datasets`. So, it may need to be changed in the example code too.
