### 2020-01-29 #2014 @lamhm
I have only four minor issues with this publication: `Performance`, `Example usage`, `Community guidelines`, and `Quality of writing`. Most of these have been pointed out by @betteridiot.

I have checked off the `Automated tests` section in my checklist because I do not have any experience with the topic (I do not do formal development in Python).
