### 2020-01-29 #2014 @will-rowe
Brilliant - thank you very much @betteridiot and @lamhm ! These are thorough reviews and I'm sure these issues can be addressed by @babinyurii

@babinyurii - please can you take a look at these issues and respond to them. Let me know if you need any assistance.
