+++ setup.py
@@ -1,7 +1,15 @@
 from distutils.core import setup
+from os import path # added to package readme on pypi
 
+this_directory = path.abspath(path.dirname(__file__))  # added to package readme on pypi
+with open(path.join(this_directory, 'README.md'), encoding='utf-8') as f:  # added to package readme on pypi
+    long_description = f.read()  # added to package readme on pypi
+    
+    
 setup(
   name = 'recan',
+  long_description = long_description,  # added to package readme on pypi
+  long_description_content_type = "text/markdown",  # added to package readme on pypi
   packages = ['recan'],   
   version = '0.1',      
   license='MIT',        
