+++ README.md
@@ -60,7 +60,7 @@ The isolate of Ba genotype is the recombinant between the virus of C genotype an
 sim_obj.simgen(window=200, shift=50, pot_rec=1)
 ```
 
-![hbv_1](pictures/HBV_1_rec_C_B_annotated.PNG)
+![hbv_1](https://raw.githubusercontent.com/babinyurii/recan/master/pictures/HBV_1_rec_C_B_annotated.PNG)
 
 
 Potential recombinant is not shown in the plot, as the distances are calculated relative to it. The higher is the distance function (i.e. the closer to 1), the closer is the sequence to the recombinant and vice versa. 
@@ -69,7 +69,7 @@ We can see typical 'crossover' of the distances which is the indicator of the po
 
 The picture from the article is shown below. It's just turned upside down relative to our plot, and instead of distance drop we see distance rising. Here Bj 'goes away' from the genotype C, whereas Ba keeps the same distance
 
-![Ba_Bj_C](pictures/hbv_C_Bj_Ba.jpg)
+![Ba_Bj_C](https://raw.githubusercontent.com/babinyurii/recan/master/pictures/hbv_C_Bj_Ba.jpg)
 
 
 By default `simgen()` method plots the whole alignment. But after initial exploration, we can take a closer look at a particular region by passing the `region` parameter to the simgen method. We can slice the alignment by using this parameter. `region` must be a tuple or a list with two integers: the start and the end position of the alignment slice.
@@ -81,7 +81,7 @@ region = (start, end)
 sim_obj.simgen(window=200, shift=50, pot_rec=1, region=(1000, 2700))
 ```
 
-![hbv_slice_1](pictures/hbv_slice_1.png)
+![hbv_slice_1](https://raw.githubusercontent.com/babinyurii/recan/master/pictures/hbv_slice_1.png)
 
 
 To customize the plot or just to export and store the data, use `get_data()` method. `get_data()` returns pandas DataFrame object with sequences as samples, and distances at given points as features. 
@@ -89,7 +89,7 @@ To customize the plot or just to export and store the data, use `get_data()` met
 ```python
 sim_obj.get_data()
 ```
-![hbv_df_example](pictures/hbv_df_example.png)
+![hbv_df_example](https://raw.githubusercontent.com/babinyurii/recan/master/pictures/hbv_df_example.png)
 
 If optional paremeter `df` is set to `False`, `get_data()` returns a tuple containing list of ticks and a dictionary of lists. Each dictionary key is the sequence id, and lists under the keys contain the corresponding distances.
 
@@ -133,7 +133,7 @@ plt.legend(prop={"size":20})
 plt.show()
 ```
 
-![hbv_matplotlib](pictures/hbv_matplotlib.png)
+![hbv_matplotlib](https://raw.githubusercontent.com/babinyurii/recan/master/pictures/hbv_matplotlib.png)
 
 
 `simgen()` method has optional parameter `dist` which denoted method used to calculate pairwise distance. By default its value is set to `pdist`, so `simgen()` calculates simple pairwise distance. To use Kimura 2 parameter distance set the value of this parameter to `k2p`
@@ -149,22 +149,22 @@ sim_obj.save_data(out="excel", out_name="hbv_distance_data")
 ```
 If there are about 20 or 30 sequences in the input file and their names are long, legend element may hide the plot. So, to be able to analyze many sequences at once, it's better to use short consice sequence names instead of long ones. Like this:
 
-![hbv_short_names](pictures/short_names.png)
[…]
