+++ setup.py
@@ -1,4 +1,3 @@
-#from distutils.core import setup
 from os import path
 from distutils.core import setup
 import setuptools 
