### 2020-02-13 babinyurii/recan#3 @babinyurii
CONTRIBUTING.md, PULL_REQUEST_TEMPLATE.md and Issue templates are added
https://github.com/babinyurii/recan
