### 2020-02-13 babinyurii/recan#4 @babinyurii
Link to the description how to install the Jupyter Lab Plotly renderer is added to the **Installation** section of the README file
https://github.com/babinyurii/recan
