### 2020-02-13 babinyurii/recan#5 @babinyurii
README file is added to the PyPI: https://pypi.org/project/recan/
