### 2020-02-13 babinyurii/recan#6 @babinyurii
The first screenshot showing similarity plot is annotated : https://raw.githubusercontent.com/babinyurii/recan/master/pictures/HBV_1_rec_C_B_annotated.PNG

Link to the dataset folder is added to README (right before the examples), and the statement `sim_obj = Simgen("./data/hbv_C_Bj_Ba.fasta` is changed to `sim_obj = Simgen("./datasets/hbv_C_Bj_Ba.fasta`
https://github.com/babinyurii/recan
