### 2020-02-13 babinyurii/recan#10 @babinyurii
I'm going to add a function which will go through the alignment and predict recombination break points based on genetic distance change. However, I need to collect and analyze more information on the confirmed recombination cases to set reasonable distance change parameters that seem to be different for RNA and DNA viruses.
Maybe it'll be better now to rename the paper to "Recan: Python tool for **analysis** of recombination events in viral genomes", as the package has no this function.
