+++ paper/paper.md
@@ -18,25 +18,22 @@ bibliography: paper.bib
 
 # Summary
 
-Recombination is widespread in viruses and is considered to  be one of the mechanisms that drives evolutionary changes and helps viruses to overcome selective pressure and adapt to new environments [@Perez-Losada2015].
-The similarity plot based on genetic distance between nucleotide sequences is an intuitive and simple way to explore and discover recombination events in viral genomes. 
-Python package named recan (recombination analyzer) allows to construct genetic distance plots and explore them interactively.  Recan is intended to be used via Jupyter notebook which is a popular instrument for data analysis.
-The aim of the package is to give the researchers who are familiar with Python a tool which allows to perform recombination analysis easily and quickly.
-The advantages of recan above the distance methods implemented previously in desktop software [@Lole1999; @Etherington2005a; @Martin2015] are its interactivity, speed, the ability to easily include sequences into or exclude them from the output without reconstructing the plot and recalculating the distance data, and, at last, the ability to analyze simultaneously several datasets during a single session.
-Recan is based on Biopython, Pandas, and Plotly libraries. 
-The package requires sequence alignment in fasta format as an input. A user can control the sliding window size, the window shift, a method of distance calculation, a sequence of interest (which is a sequence potentially containing recombination region), and the alignment region which will be included into distance calculation. Two methods of genetic distance calculation implemented so far are pairwise and Kimura 2-parameter distance.  After a recombination event is found, the distance data can be saved in csv or excel  file, or directly used to reconstruct the plot in Jupyter notebook with ones favorite plotting library to get a final report.
+Recombination drives virus evolution in response to  selective forces in a host enviroment and adaptation to new abiotoic factors  [@Perez-Losada2015].
+Gaining insights into such genetic events is important for a better understanding of viral biology.  This task could be achieved through analysis of similarity plots based on a calculating genetic distances between nucleotide sequences.
+Python package named «recan» (recombination analyzer) provides the means to construct genetic distance plots and explore them interactively. The package has been designed to operate via the Jupyter notebook. Recan is an amenable tool able to help scientists who are experienced in Python to analyze recombination events.
+
+Compared to the previously designed desktop software  [@Lole1999; @Etherington2005a; @Martin2015] recan has the ability to insert or delete sequences from the output without reconstructing the plots and recalculating the distance values. Finally, recan enables simultaneous analysis of several datasets in a single session. Recan is based on Biopython, Pandas, and Plotly libraries. The package requires a sequence alignment in fasta format as an input. The user can adjust the sliding window size, the window shift, method of distance calculation,  sequence of interest (a region where breakpoints occur ), and the length of alignment region which will be included into the distance calculation . The two methods of genetic distance calculations implemented in recan are the pairwise and Kimura 2-parameter models. The distance data can be saved in csv or excel file, or directly used to reconstruct the plot in Jupyter notebook in the preferable plotting library to obtain a final report. 
 
 # Testing and verification
-To test the package four previously reported recombinant viral genomes were used: human immunodeficiency virus (HIV) [@Liitsola2000a], hepatitis C virus (HCV) [@Smith2014] norovirus [@Jiang1999], and lumpy skin disease virus (LSDV) [@Sprygin2018].
-Each dataset included a recombinant sequence, its parental sequences and a number of other sequences of the same virus closely related to the recombinant.
[…]
