+++ paper/paper.md
@@ -20,17 +20,17 @@ bibliography: paper.bib
 
 Recombination drives virus evolution in response to  selective forces in a host enviroment and adaptation to new abiotoic factors  [@Perez-Losada2015].
 Gaining insights into such genetic events is important for a better understanding of viral biology.  This task could be achieved through analysis of similarity plots based on a calculating genetic distances between nucleotide sequences.
-Python package named «recan» (recombination analyzer) provides the means to construct genetic distance plots and explore them interactively. The package has been designed to operate via the Jupyter notebook. Recan is an amenable tool able to help scientists who are experienced in Python to analyze recombination events.
+Python package named "recan" (recombination analyzer) provides the means to construct genetic distance plots and explore them interactively. The package has been designed to operate via the Jupyter notebook. Recan is an amenable tool able to help scientists who are experienced in Python to analyze recombination events.
 
 Compared to the previously designed desktop software  [@Lole1999; @Etherington2005a; @Martin2015] recan has the ability to insert or delete sequences from the output without reconstructing the plots and recalculating the distance values. Finally, recan enables simultaneous analysis of several datasets in a single session. Recan is based on Biopython, Pandas, and Plotly libraries. The package requires a sequence alignment in fasta format as an input. The user can adjust the sliding window size, the window shift, method of distance calculation,  sequence of interest (a region where breakpoints occur ), and the length of alignment region which will be included into the distance calculation . The two methods of genetic distance calculations implemented in recan are the pairwise and Kimura 2-parameter models. The distance data can be saved in csv or excel file, or directly used to reconstruct the plot in Jupyter notebook in the preferable plotting library to obtain a final report. 
 
 # Testing and verification
 To test the package, we used four previously reported recombinant viral genomes representing different genuses: human immunodeficiency virus (HIV) [@Liitsola2000a],hepatitis C virus (HCV)  [@Smith2014] norovirus [@Jiang1999], and lumpy skin disease virus (LSDV)  [@Sprygin2018].
 Each dataset included a recombinant virus sequence, its putative parental sequences and a set of sequences of the same virus closely related to the analysed recombinant virus. HIV, HCV and Norovirus sequences were aligned using  Clustal W  [@Larkin2007] and LSDV genomes were aligned using MAFFT [@Katoh2002] as part of Ugene software  [@Okonechnikov2012].
-The HIV alignment contained  twenty five 3135 bp sequences ; The HCV alignment contained twenty three 9431 bp sequences ; The norovirus alignment included nineteen 3366 bp sequecnes, and the  LSDV alignment had a total of 150511 bp  sequences . The resulting simgen method execution time with the default window size and shift parameters was the following: 437 ms ± 7.74 ms for HIV, 579 ms ± 58.7 ms for Norovirus, 648 ms ± 44.2 ms for HCV, and 3.55 s ± 239 ms for LSDV dataset. Time execution test was performed using a desktop PC with 4 CPU cores and 4 Gb RAM. LSDV has one of the largest genomes of all viruses (about 150 000 bp). Ultimately, recan can potentially identify recombination events in a large subset of sequences regardless of the length of the viral genome. The distance plots with recombination events detected by recan are shown in Figures 1-4.
[…]
