+++ paper/paper.md
@@ -18,16 +18,15 @@ bibliography: paper.bib
 
 # Summary
 
-Recombination drives virus evolution in response to  selective forces in a host enviroment and adaptation to new abiotoic factors  [@Perez-Losada2015].
-Gaining insights into such genetic events is important for a better understanding of viral biology.  This task could be achieved through analysis of similarity plots based on a calculating genetic distances between nucleotide sequences.
+Recombination drives virus evolution in response to selective forces in a host environment and adaptation to new abiotic factors [@Perez-Losada2015].
+Gaining insights into such genetic events is important for a better understanding of viral biology. This task could be achieved through analysis of similarity plots based on a calculating genetic distances between nucleotide sequences.
 Python package named "recan" (recombination analyzer) provides the means to construct genetic distance plots and explore them interactively. The package has been designed to operate via the Jupyter notebook. Recan is an amenable tool able to help scientists who are experienced in Python to analyze recombination events.
-
-Compared to the previously designed desktop software  [@Lole1999; @Etherington2005a; @Martin2015] recan has the ability to insert or delete sequences from the output without reconstructing the plots and recalculating the distance values. Finally, recan enables simultaneous analysis of several datasets in a single session. Recan is based on Biopython, Pandas, and Plotly libraries. The package requires a sequence alignment in fasta format as an input. The user can adjust the sliding window size, the window shift, method of distance calculation,  sequence of interest (a region where breakpoints occur ), and the length of alignment region which will be included into the distance calculation . The two methods of genetic distance calculations implemented in recan are the pairwise and Kimura 2-parameter models. The distance data can be saved in csv or excel file, or directly used to reconstruct the plot in Jupyter notebook in the preferable plotting library to obtain a final report. 
+Compared to the previously designed desktop software [@Lole1999; @Etherington2005a; @Martin2015] recan has the ability to insert or delete sequences from the output without reconstructing the plots and recalculating the distance values. Finally, recan enables simultaneous analysis of several datasets in a single session. Recan is based on Biopython, Pandas, and Plotly libraries. The package requires a sequence alignment in fasta format as an input. The user can adjust the sliding window size, the window shift, method of distance calculation, sequence of interest (a region where breakpoints occurs), and the length of alignment region which will be included into the distance calculation . The two methods of genetic distance calculations implemented in recan are the pairwise and Kimura 2-parameter models. The distance data can be saved in csv or excel file, or directly used to reconstruct the plot in the Jupyter notebook using the preferable plotting library to obtain a final report. 
 
 # Testing and verification
-To test the package, we used four previously reported recombinant viral genomes representing different genuses: human immunodeficiency virus (HIV) [@Liitsola2000a],hepatitis C virus (HCV)  [@Smith2014] norovirus [@Jiang1999], and lumpy skin disease virus (LSDV)  [@Sprygin2018].
-Each dataset included a recombinant virus sequence, its putative parental sequences and a set of sequences of the same virus closely related to the analysed recombinant virus. HIV, HCV and Norovirus sequences were aligned using  Clustal W  [@Larkin2007] and LSDV genomes were aligned using MAFFT [@Katoh2002] as part of Ugene software  [@Okonechnikov2012].
[…]
