### 2020-02-17 3db5e4b
Update paper.md
