+++ paper/paper.md
@@ -25,7 +25,7 @@ Compared to the previously designed desktop software [@Lole1999; @Etherington200
 
 # Testing and verification
 To test the package, we used four previously reported recombinant viral genomes representing different genuses: human immunodeficiency virus (HIV) [@Liitsola2000a], hepatitis C virus (HCV) [@Smith2014] norovirus [@Jiang1999], and lumpy skin disease virus (LSDV) [@Sprygin2018].
-Each dataset included a recombinant virus sequence, its putative parental sequences and a set of sequences of the same virus closely related to the recombinant virus. HIV, HCV and Norovirus sequences were aligned using ClustalW [@Larkin2007] and LSDV genomes were aligned using MAFFT [@Katoh2002] as part of Ugene software [@Okonechnikov2012].
+Each dataset included a recombinant virus sequence, its putative parental sequences and a set of sequences of the same virus closely related to the recombinant virus. HIV, HCV and Norovirus sequences were aligned using ClustalW [@Larkin2007], and LSDV genomes were aligned using MAFFT [@Katoh2002] as part of Ugene software [@Okonechnikov2012].
 The HIV alignment contained twenty five 3135 bp sequences ; the HCV alignment contained twenty three 9431 bp sequences ; the norovirus alignment included nineteen 3366 bp sequecnes, and the LSDV alignment had a total of 150511 bp sequences . The resulting `simgen` method execution time with the default window size and shift parameters was the following: 437 ms ± 7.74 ms for HIV, 579 ms ± 58.7 ms for Norovirus, 648 ms ± 44.2 ms for HCV, and 3.55 s ± 239 ms for LSDV dataset. Time execution test was performed using a desktop PC with 4 CPU cores and 4 Gb RAM. LSDV has one of the largest genomes of all viruses (about 150 000 bp). Ultimately, recan can potentially be used to identify and analize recombination events in a large subset of sequences regardless of the length of the viral genome. The distance plots with recombination events detected by recan are shown in Figures 1-4.
 
 # Availability and implementation
