+++ README.md
@@ -172,6 +172,7 @@ To verify the installation, go to the `recan/test/` folder and run:
 `
 $ pytest test.py
 `
+
 or
 
 `
@@ -179,7 +180,7 @@ $ python -m unittest test.py
 `
 
 
-**references**
+## References
 
 1. Recombination Analysis Tool (RAT): a program for the high-throughput detection of recombination. Bioinformatics, Volume 21, Issue 3,
 1 February 2005, Pages 278–281, https://doi.org/10.1093/bioinformatics/bth500
