### 2020-02-17 babinyurii/recan#7 @babinyurii
1. and 3. instruction how to run the tests is added to the README
https://github.com/babinyurii/recan
2. folder `test` is renamed to `tests` and `test.py` to `basic_test.py`.
4. I think the coverage is rather low because the large part of code just creates plots. I can't test the plotting statements themselves, and thus the test verifies input genetic distance values for plotting methods.
