### 2020-02-17 babinyurii/recan#11 @babinyurii
The paper has been edited. Perhaps, the language is a bit better now.
