### 2020-02-18 #2014 @will-rowe
great - thanks @babinyurii

Could you please respond to [this reviewer issue](https://github.com/babinyurii/recan/issues/12).

You also suggested two options in response to [this issue](https://github.com/babinyurii/recan/issues/10). Can you please let us know which you would like to proceed with? I.E. renaming the paper and adjusting the scope, or the addition of a function for detection of recombination breakpoints
