### 2020-02-19 6a3f01f
Update paper.md
