### 2020-02-19 babinyurii/recan#12 @babinyurii
Thank you for valuable comment, @lamhm. I agree with you, the speed comparison was incorrect. Now the paper was edited, and that statement was deleted.
