### 2020-02-19 babinyurii/recan#10 @babinyurii
I'd like to proceed the review process with renaming the article to "Recan: Python tool for **analysis** of recombination events in viral genomes".
https://github.com/babinyurii/recan/blob/master/paper/paper.md
