### 2020-02-19 #2014 @will-rowe
Great - thanks @babinyurii

I think we're getting close here. Can I now ask @betteridiot and @lamhm to review the author responses to the issues raised and check they are satisfied? From my point of view, I am happy now that @babinyurii has adjusted the scope of the paper to focus on the analysis, rather than detection, of recombination events.
