### 2020-03-02 #2014 @betteridiot
Sorry. I might be able to get to it later this week. I am out of town at a conference. Kind of spotty internet at best.

Marcus D. Sherman
RISE & RMF Fellow
PhD Candidate
Mills Lab
Dept. of Comp. Medicine & Bioinformatics
University of Michigan
https://twitter.com/better_idiot
https://github.com/betteridiot

From: Will Rowe
Sent: Monday, March 2, 2020 5:13 AM
To: openjournals/joss-reviews
Cc: Marcus D Sherman; Mention
Subject: Re: [openjournals/joss-reviews] [REVIEW]: Recan: Python tool fordetection of recombination events in viral genomes (#2014)

Pinging @betteridiot and @lamhm - could I ask you to take another look at this submission and see if you are happy that your comments have been addressed to your satisfaction?
