### 2020-03-03 babinyurii/recan#12 @ghost
Thanks a lot, @babinyurii. I confirm that the performance statement has been corrected in the paper. I will close this issue now.
