### 2020-03-03 babinyurii/recan#10 @ghost
I understand that developing a method for recombination detection is not easy and will take some time. Perhaps, you may want to consider incorporating a currently existing method into your package.

For the moment, I think changing the title of the paper is good enough.
