### 2020-03-03 babinyurii/recan#11 @ghost
Hi @babinyurii,

There seems to be a missing piece between the 2nd and the 3rd sentences in your summary.

> Gaining insights into such genetic events is important for a better understanding
of viral biology. This task could be achieved through analysis of similarity plots based on
calculating genetic distances between nucleotide sequences.

I think you meant to go from `recombination events` to `recombination detection analysis` to `similarity plots`. But here, the middle link is missing. So, the logic sounds a bit odd.

I can tell there are still a few grammar mistakes in the manuscript. E.g. the article "the" is missing here and there. But I am not a native English speaker. Perhaps @betteridiot could give you more helpful comments for this.
