### 2020-03-03 #2014 @lamhm
Hi @babinyurii and @will-rowe,

Apologies for my delayed response. I think `recan` has fulfilled most of the criteria in the review checklist. I only have two comments left to be resolved. They are in the `Example usage` and the `Quality of writing` issues.
