### 2020-03-12 babinyurii/recan#6 @babinyurii
Thanks for comment, @lamhm. Now I've added the link to download the example datasets. It's in the section `example datasets` which is right above the `references` section.
