### 2020-03-13 #2014 @will-rowe
That's great @babinyurii. When @betteridiot signs off on his review, I will go through and help edit the paper.

Pinging @betteridiot - I know it's hectic times at the moment and I understand if this has dropped off your radar. Please let us know if you'll be able to check your comments have been addressed. If you can't manage it in the next week or two, I'll make the decision based on what you have provided so far.

Cheers,

Will
