### 2020-03-20 babinyurii/recan#6 @ghost
Thanks a lot, @babinyurii. The example usage looks great now.

@betteridiot - If you do not have any additional concerns, I think you can close this issue.
