### 2020-04-03 #2014 @betteridiot
I apologize for the long delay in responding. The author has addressed most of my notes and looks to be actively working on the "quality of writing". With that, I believe the paper is ready to move forward.
