### 2020-04-13 2bbd181
Update paper.md
* changed constructing to construction to match the tense of the sentence.
* removed unnecessary sentence that did not make clear sense:
 `Recan is an amenable tool able to help scientists who are familiar with Python to analyze recombination events.`
* changed `via the Jupyter notebook` to ` with Jupyter notebooks`
* removed the word preferable, which made the sentence hard to read
