### 2020-04-13 #2014 @will-rowe
Sorry for the delay @babinyurii (and @arfon) - I've been super busy with COVID stuff and my editorial responsibilities have unfortunately been neglected. My apologies.

The paper is looking great - thank you for you work in addressing all the reviewer comments and for improving the writing quality.

I have just submitted a small pull request to your repo that makes a few grammatically changes. https://github.com/babinyurii/recan/pull/15 - let me know if you agree with these. Once you are happy, I think we can move to acceptance here.
