### 2020-05-09 #2014 @babinyurii
Hi @will-rowe! Thank you for reminding me. I'm sorry for overlooking your merge in a hurry. I've already put the changes you proposed into the paper.
